refactor: Extract shared dither spec helpers into DitherSpecHelpers module
Deduplicates build_gradient_canvas, render_with, count_transitions,
and collect_pixels from atkinson_spec and floyd_steinberg_spec into a
shared module in shared_examples.rb.
