+++ spec/chroma_wave/dither/atkinson_spec.rb
@@ -3,6 +3,8 @@
 require_relative 'shared_examples'
 
 RSpec.describe ChromaWave::Dither::Atkinson do
+  include DitherSpecHelpers
+
   let(:mono_format)  { ChromaWave::PixelFormat::MONO }
   let(:gray_format)  { ChromaWave::PixelFormat::GRAY4 }
   let(:color4_format) { ChromaWave::PixelFormat::COLOR4 }
@@ -77,43 +79,5 @@ RSpec.describe ChromaWave::Dither::Atkinson do
       strategy = described_class.new(pixel_format: mono_format)
       expect { strategy.call(canvas, framebuffer) }.not_to raise_error
     end
-  end
-
-  private
-
-  # Builds a horizontal gradient canvas from black to white.
-  def build_gradient_canvas(width:, height:)
-    canvas = ChromaWave::Canvas.new(width: width, height: height, background: ChromaWave::Color::BLACK)
-    width.times do |x|
-      val = (x * 255.0 / (width - 1)).round
-      color = ChromaWave::Color.new(r: val, g: val, b: val)
-      height.times { |y| canvas.set_pixel(x, y, color) }
-    end
-    canvas
-  end
-
-  # Renders a canvas using a strategy class and format.
-  def render_with(strategy_class, format, canvas)
-    framebuffer = ChromaWave::Framebuffer.new(canvas.width, canvas.height, format)
-    strategy_class.new(pixel_format: format).call(canvas, framebuffer)
-    framebuffer
-  end
-
-  # Counts color transitions across a single-row framebuffer.
-  def count_transitions(framebuffer, width)
-    transitions = 0
-    (1...width).each do |x|
-      transitions += 1 if framebuffer.get_pixel(x, 0) != framebuffer.get_pixel(x - 1, 0)
-    end
-    transitions
-  end
-
-  # Collects all pixel values from a framebuffer into a flat array.
-  def collect_pixels(framebuffer, width, height)
-    pixels = []
-    height.times do |y|
-      width.times { |x| pixels << framebuffer.get_pixel(x, y) }
-    end
-    pixels
   end
 end
